start to move object creation routines to factory.c
