+++ src/express/symbol.c
@@ -39,16 +39,3 @@
 void SYMBOLinitialize( void ) {
     ALLOCinitialize( &SYMBOL_fl, sizeof( struct Symbol_ ), 100, 100 );
 }
-
-Symbol * SYMBOLcreate( char * name, int line, const char * filename ) {
-    Symbol * sym = SYMBOL_new();
-    sym->name = name;
-    sym->line = line;
-    sym->filename = filename; /* NOTE this used the global 'current_filename',
-                               * instead of 'filename'. This func is only
-                               * called in two places, and both calls use
-                               * 'current_filename'. Changed this to avoid
-                               * potential future headaches. (MAP, Jan 12)
-                               */
-    return sym;
-}
