+++ src/express/type.c
@@ -162,33 +162,6 @@ Type Type_Bag_Of_Generic;
 Error ERROR_corrupted_type = ERROR_none;
 
 static Error ERROR_undefined_tag;
-/**
- * create a type with no symbol table
- */
-Type TYPEcreate_nostab( struct Symbol_ *symbol, Scope scope, char objtype ) {
-    Type t = SCOPEcreate_nostab( OBJ_TYPE );
-    TypeHead th = TYPEHEAD_new();
-
-    t->u.type = th;
-    t->symbol = *symbol;
-    DICTdefine( scope->symbol_table, symbol->name, ( Generic )t, &t->symbol, objtype );
-
-    return t;
-}
-
-/**
- * create a type but this is just a shell, either to be completed later
- * such as enumerations (which have a symbol table added later)
- * or to be used as a type reference
- */
-Type TYPEcreate_name( Symbol * symbol ) {
-    Scope s = SCOPEcreate_nostab( OBJ_TYPE );
-    TypeHead t = TYPEHEAD_new();
-
-    s->u.type = t;
-    s->symbol = *symbol;
-    return s;
-}
 
 Type TYPEcreate_user_defined_tag( Type base, Scope scope, struct Symbol_ *symbol ) {
     Type t;
@@ -227,29 +200,6 @@ Type TYPEcreate_user_defined_tag( Type base, Scope scope, struct Symbol_ *symbol
     tag_count++;
 
     return( t );
-}
-
-Type TYPEcreate( enum type_enum type ) {
-    TypeBody tb = TYPEBODYcreate( type );
-    Type t = TYPEcreate_from_body_anonymously( tb );
-    return( t );
-}
-
-Type TYPEcreate_from_body_anonymously( TypeBody tb ) {
-    Type t = SCOPEcreate_nostab( OBJ_TYPE );
-    TypeHead th = TYPEHEAD_new();
-
-    t->u.type = th;
-    t->u.type->body = tb;
-    t->symbol.name = 0;
-    SYMBOLset( t );
-    return t;
-}
-
-TypeBody TYPEBODYcreate( enum type_enum type ) {
-    TypeBody tb = TYPEBODY_new();
-    tb->type = type;
-    return tb;
 }
 
 /**
